software/ground/requirements.txt
pyserial
numpy
matplotlib
pyqt6